fixed outright null value causing crash. / updated values & added null check.

See merge request Android/sport-android!652

+++ app/src/main/java/org/cxct/sportlottery/ui/odds/OddsDetailFragment.kt
@@ -221,8 +221,8 @@ class OddsDetailFragment : BaseSocketFragment<GameViewModel>(GameViewModel::clas
 
     private fun setupStartTime(matchInfo: MatchInfo?) {
         matchInfo?.apply {
-            tv_time_top.text = TimeUtil.timeFormat(startTime.toLong(), TimeUtil.DM_FORMAT)
-            tv_time_bottom.text = TimeUtil.timeFormat(startTime.toLong(), TimeUtil.HM_FORMAT)
+            tv_time_top.text = TimeUtil.timeFormat(startTime, TimeUtil.DM_FORMAT)
+            tv_time_bottom.text = TimeUtil.timeFormat(startTime, TimeUtil.HM_FORMAT)
         }
     }
